requirements.txt
django
djangorestframework==3.13.1
mysql-connector==2.2.9
mysqlclient==2.1.0
pymongo==4.0.1
python-dateutil==2.8.2